Default values of cfg.get changes to match expected type
Due to template argument deduction, if the default value (second
argument) does not match the type of the variable (e.g., 2 vs 2.0) there
are precision misses if the value from file has a nonzero decimal value.

For example, if the default of wind_height is 2 but the value from file
is 8.6 the 0.6 will be lost and only 8.0 will be retained.