Correcting `base_step` constness, concepts, and constructors
**Constness**

`GuaranteeImplementExecute` now accepts `const T&` and `Data&` types.
This ensures that (a) `T` has a function `void execute_impl` that
takes a `Data&` and (b) `T` can be `const` and or a reference.

**Concepts**

The previous concept was circular: a derived class would need to be
declared like:

```cpp
template<typename Data>
class Derived : public base_step<Derived,Data>
```

But since `Derived` is not fully defined, it cannot satisfy the concept.
Instead, this is replaced with a static assert that is only tripped if
`execute` is called. It is still tripped at compile time and therefore
has the same effect.

**Constructors**

The constructor is now explicitly declared and is protected.